refactor: move people sorting into src/lib/sorting/people.ts
- Created `src/lib/sorting/people.ts` with 6 reusable sorting/grouping functions
- Replaced inline sorting logic in PartyCard.tsx and chat route
- Updated imports across 6 files to use centralized module
- Deleted `src/components/utils.ts`
