+++ src/lib/sorting/people.ts
@@ -151,3 +151,4 @@ export const sortPeople = (
 
     return sortedPeople;
 };
+
